AIML ROC_AUC_Task_day13
I performed this task in Pycharm and Juypter notebook in this task I imported libraries like Sklearn and matplotlib .In this first I imported libraries, load datasets, spilt it to train and and test model model, scale the data, train the logistic Regression model, predict probabilities  and lastly calculate AUC score and plot ROC curve. While performing this task I understand and learn more about ROC curve and AUC  value. I learnt that ROC is a curve of true positive rate and false positive rate graph and AUC is area of that curve