Numpy_advanced Day 22
I performed this task in Juypter notebook with the help of library Numpy .While Performing This task I understand the use of this library that is it help us to perform various task in arrays. In this task I make one matrix with the help of this library and then Squared it, Vectorized  it ,slice it and multiply this matrix to another. Overall this task is very easy